scripts/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.0
markdown>=3.5.0
PyGithub>=2.1.1
lxml>=4.9.0
html2text>=2020.1.16
